added start script to backend and some file structure, fixed getopenRoles try-catch
