+++ backend/package.json
@@ -1,9 +1,10 @@
 {
-  "name": "backend",
+  "name": "community-bulletin-backend",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/http.js",
   "scripts": {
+    "start": "nodemon src/http.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -11,6 +12,7 @@
   "dependencies": {
     "airtable": "^0.10.1",
     "dotenv": "^8.2.0",
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "nodemon": "^2.0.6"
   }
 }
